Removed NewDocumentService, moved contents to DocumentService
