+++ packaging/wcm/webapp/src/main/webapp/WEB-INF/conf/wcm-extension/wcm/new-document-configuration.xml
@@ -1,11 +1,6 @@
 <configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd"
   xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
-
-  <component>
-    <key>org.exoplatform.ecm.webui.component.explorer.documents.NewDocumentService</key>
-    <type>org.exoplatform.ecm.webui.component.explorer.documents.NewDocumentServiceImpl</type>
-  </component>
   
     <!-- ECMS menu action for New Document. -->
   <external-component-plugins>
